+++ packages/benchmark-js/src/Benchmark.ts
@@ -1,11 +1,12 @@
 import { tTable } from './constants';
-import { TimeUnitUtil } from './TimeUnitUtil';
+import { Logger, LoggerLevel } from './Logger';
+import { TU } from './TimeUnit';
 import { BenchmarkOptions, TestFn } from './types';
 import { BenchmarkStats, _BenchmarkSettings, _Nanosecond } from './types.internal';
 import { formatNumber, genStr, getMean, getMinTime, getVariance, hrtime2ns, sleep } from './utils';
 
 export class Benchmark {
-    private static minTime: _Nanosecond = 0 as _Nanosecond;
+    private static minTime: _Nanosecond = TU.ns(0);
 
     private name: string;
 
@@ -18,7 +19,7 @@ export class Benchmark {
 
     private stats: BenchmarkStats = {
         deviation: 0,
-        mean: 0 as _Nanosecond,
+        mean: TU.ns(0),
         moe: 0,
         ops: 0,
         rme: 0,
@@ -50,22 +51,28 @@ export class Benchmark {
         this.onStart = options?.onStart ?? null;
 
         if (Benchmark.minTime === 0) {
-            Benchmark.minTime = Math.max(getMinTime() * 100, 50_000_000) as _Nanosecond;
+            Benchmark.minTime = TU.ns(Math.max(getMinTime() * 100, 50_000_000));
         }
 
         this.settings = {
-            delay: TimeUnitUtil.ms2ns(options?.delay ?? 5),
+            delay: TU.ms2ns(options?.delay ?? 5),
             initCount: options?.initCount ?? 1,
-            maxPrepareTime: TimeUnitUtil.ms2ns(options?.maxPrepareTime ?? 1_000),
-            maxTime: TimeUnitUtil.ms2ns(options?.maxTime ?? 5_000),
+            maxPrepareTime: TU.ms2ns(options?.maxPrepareTime ?? 1_000),
+            maxTime: TU.ms2ns(options?.maxTime ?? 5_000),
             minSamples: options?.minSamples ?? 5,
             minTime:
-                options?.minTime === undefined || options.minTime === 0
-                    ? Benchmark.minTime
-                    : TimeUnitUtil.ms2ns(options.minTime),
+                options?.minTime === undefined || options.minTime === 0 ? Benchmark.minTime : TU.ms2ns(options.minTime),
         };
 
         this.count = this.settings.initCount;
+    }
+
+    public static get loggerLevel(): LoggerLevel {
+        return Logger.level;
+    }
+
+    public static set loggerLevel(level: LoggerLevel) {
+        Logger.level = level;
     }
 
     /**
@@ -98,10 +105,10 @@ export class Benchmark {
     }
 
     private benchmarking(ns: _Nanosecond): void {
-        let elapsed: _Nanosecond = 0 as _Nanosecond;
+        let elapsed: _Nanosecond = TU.ns(0);
         while (true) {
             const used = this.cycle();
-            const period = (used / this.count) as _Nanosecond;
+            const period = TU.ns(used / this.count);
             this.stats.sample.push(period);
 
             // Calculate how many more iterations it will take to achieve the `minTime`.
@@ -110,7 +117,7 @@ export class Benchmark {
                 this.count += Math.ceil((this.settings.minTime - used) / period);
             }
 
-            elapsed = (elapsed + used) as _Nanosecond;
+            elapsed = TU.ns(elapsed + used);
             if (elapsed >= ns && this.stats.sample.length >= this.settings.minSamples) break;
 
             sleep(this.settings.delay);
@@ -159,6 +166,6 @@ export class Benchmark {
     }
 
     public printResult(): void {
-        console.info(this.toString());
+        Logger.log(this.toString());
     }
 }