+++ packages/benchmark-js/src/index.ts
@@ -1,2 +1,3 @@
 export * from './Benchmark';
 export * from './BenchmarkGroup';
+export { LoggerLevel } from './Logger';